roachtest: move replication functions to `roachtestutil`
This moves replication-related functions (such as
`WaitForReplication`) from the `util.go` file in roachtest to the
`roachtestutil` package. This will allow this functionality to be
reused in roachtest itself.

Epic: none

Release note: None
