+++ pkg/cmd/roachtest/roachtestutil/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "health_checker.go",
         "httpclient.go",
         "jaeger.go",
+        "replication.go",
         "utils.go",
         "validation_check.go",
     ],
@@ -28,6 +29,7 @@ go_library(
         "//pkg/util/humanizeutil",
         "//pkg/util/log",
         "//pkg/util/protoutil",
+        "//pkg/util/retry",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
