+++ pkg/cmd/roachtest/tests/cdc_bench.go
@@ -267,7 +267,7 @@ func runCDCBenchScan(
 	}
 
 	// Wait for system ranges to upreplicate.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create and split the workload table. We don't import data here, because it
 	// imports before splitting, which takes a very long time.
@@ -277,7 +277,7 @@ func runCDCBenchScan(
 	t.L().Printf("creating table with %s ranges", humanize.Comma(numRanges))
 	c.Run(ctx, option.WithNodes(nCoord), fmt.Sprintf(
 		`./cockroach workload init kv --splits %d {pgurl:%d}`, numRanges, nData[0]))
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	cursor := timeutil.Now() // before data is ingested
 
@@ -445,7 +445,7 @@ func runCDCBenchWorkload(
 	}
 
 	// Wait for system ranges to upreplicate.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create and split the workload table.
 	//
@@ -454,7 +454,7 @@ func runCDCBenchWorkload(
 	t.L().Printf("creating table with %s ranges", humanize.Comma(numRanges))
 	c.Run(ctx, option.WithNodes(nWorkload), fmt.Sprintf(
 		`./cockroach workload init kv --splits %d {pgurl:%d}`, numRanges, nData[0]))
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// For read-only workloads, ingest some data. init --insert-count does not use
 	// the standard key generator that the read workload uses, so we have to write
