+++ pkg/cmd/roachtest/tests/cluster_init.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
 	"github.com/cockroachdb/cockroach/pkg/server/authserver"
@@ -156,7 +157,7 @@ func runClusterInit(ctx context.Context, t test.Test, c cluster.Cluster) {
 		c.Run(ctx, option.WithNodes(c.Node(initNode)), `./cockroach init --url={pgurl:1}`)
 
 		// This will only succeed if 3 nodes joined the cluster.
-		err = WaitFor3XReplication(ctx, t, t.L(), dbs[0])
+		err = roachtestutil.WaitFor3XReplication(ctx, t, t.L(), dbs[0])
 		require.NoError(t, err)
 
 		execCLI := func(runNode int, extraArgs ...string) (string, error) {
