+++ pkg/cmd/roachtest/tests/decommissionbench.go
@@ -400,7 +400,7 @@ func setupDecommissionBench(
 		// import can saturate snapshots and leave underreplicated system ranges
 		// struggling.
 		// See GH issue #101532 for longer term solution.
-		if err := WaitForReplication(ctx, t, t.L(), db, 3, atLeastReplicationFactor); err != nil {
+		if err := roachtestutil.WaitForReplication(ctx, t, t.L(), db, 3, roachtestutil.AtLeastReplicationFactor); err != nil {
 			t.Fatal(err)
 		}
 
@@ -439,7 +439,7 @@ func setupDecommissionBench(
 		}
 
 		// Wait for initial up-replication.
-		err := WaitFor3XReplication(ctx, t, t.L(), db)
+		err := roachtestutil.WaitFor3XReplication(ctx, t, t.L(), db)
 		require.NoError(t, err)
 	}
 }
