+++ pkg/cmd/roachtest/tests/failover.go
@@ -241,7 +241,7 @@ func runFailoverChaos(ctx context.Context, t test.Test, c cluster.Cluster, readO
 
 	// Wait for upreplication.
 	require.NoError(
-		t, WaitForReplication(ctx, t, t.L(), conn, 5 /* replicationFactor */, atLeastReplicationFactor),
+		t, roachtestutil.WaitForReplication(ctx, t, t.L(), conn, 5 /* replicationFactor */, roachtestutil.AtLeastReplicationFactor),
 	)
 
 	// Create the kv database. If this is a read-only workload, populate it with
@@ -264,7 +264,7 @@ func runFailoverChaos(ctx context.Context, t test.Test, c cluster.Cluster, readO
 
 	// Wait for upreplication of the new ranges.
 	require.NoError(
-		t, WaitForReplication(ctx, t, t.L(), conn, 5 /* replicationFactor */, atLeastReplicationFactor),
+		t, roachtestutil.WaitForReplication(ctx, t, t.L(), conn, 5 /* replicationFactor */, roachtestutil.AtLeastReplicationFactor),
 	)
 
 	// Run workload on n10 via n1-n2 gateways until test ends (context cancels).
@@ -425,7 +425,7 @@ func runFailoverPartialLeaseGateway(ctx context.Context, t test.Test, c cluster.
 	configureAllZones(t, ctx, conn, zoneConfig{replicas: 3, onlyNodes: []int{1, 2, 3}})
 
 	// Wait for upreplication.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create the kv database with 5 replicas on n2-n6, and leases on n4.
 	t.L().Printf("creating workload database")
@@ -558,8 +558,8 @@ func runFailoverPartialLeaseLeader(ctx context.Context, t test.Test, c cluster.C
 	configureAllZones(t, ctx, conn, zoneConfig{replicas: 3, onlyNodes: []int{1, 2, 3}})
 
 	// NB: We want to ensure the system ranges are all down-replicated from their
-	// initial RF of 5, so pass in exactlyReplicationFactor below.
-	require.NoError(t, WaitForReplication(ctx, t, t.L(), conn, 3, exactlyReplicationFactor))
+	// initial RF of 5, so pass in roachtestutil.ExactlyReplicationFactor below.
+	require.NoError(t, roachtestutil.WaitForReplication(ctx, t, t.L(), conn, 3, roachtestutil.ExactlyReplicationFactor))
 
 	// Now that system ranges are properly placed on n1-n3, start n4-n6.
 	c.Start(ctx, t.L(), failoverStartOpts(), settings, c.Range(4, 6))
@@ -681,7 +681,7 @@ func runFailoverPartialLeaseLiveness(ctx context.Context, t test.Test, c cluster
 		replicas: 4, onlyNodes: []int{1, 2, 3, 4}, leasePreference: "[+node4]"})
 
 	// Wait for upreplication.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create the kv database on n5-n7.
 	t.L().Printf("creating workload database")
@@ -795,7 +795,7 @@ func runFailoverNonSystem(
 	configureAllZones(t, ctx, conn, zoneConfig{replicas: 3, onlyNodes: []int{1, 2, 3}})
 
 	// Wait for upreplication.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create the kv database, constrained to n4-n6. Despite the zone config, the
 	// ranges will initially be distributed across all cluster nodes.
@@ -905,7 +905,7 @@ func runFailoverLiveness(
 	configureZone(t, ctx, conn, `RANGE liveness`, zoneConfig{replicas: 4, leasePreference: "[+node4]"})
 
 	// Wait for upreplication.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create the kv database, constrained to n1-n3. Despite the zone config, the
 	// ranges will initially be distributed across all cluster nodes.
@@ -1017,7 +1017,7 @@ func runFailoverSystemNonLiveness(
 	configureZone(t, ctx, conn, `RANGE liveness`, zoneConfig{replicas: 3, onlyNodes: []int{1, 2, 3}})
 
 	// Wait for upreplication.
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 	// Create the kv database, constrained to n1-n3. Despite the zone config, the
 	// ranges will initially be distributed across all cluster nodes.
@@ -1605,7 +1605,7 @@ func (f *pauseFailer) Recover(ctx context.Context, nodeID int) {
 // waitForUpreplication waits for upreplication of ranges that satisfy the
 // given predicate (using SHOW RANGES).
 //
-// TODO(erikgrinaker): move this into WaitForReplication() when it can use SHOW
+// TODO(erikgrinaker): move this into roachtestutil.WaitForReplication() when it can use SHOW
 // RANGES, i.e. when it's no longer needed in mixed-version tests with older
 // versions that don't have SHOW RANGES.
 func waitForUpreplication(
