+++ pkg/cmd/roachtest/tests/gossip.go
@@ -42,7 +42,7 @@ func registerGossip(r registry.Registry) {
 		startOpts := option.DefaultStartOpts()
 		startOpts.RoachprodOpts.ExtraArgs = append(startOpts.RoachprodOpts.ExtraArgs, "--vmodule=*=1")
 		c.Start(ctx, t.L(), startOpts, install.MakeClusterSettings(), c.All())
-		err := WaitFor3XReplication(ctx, t, t.L(), c.Conn(ctx, t.L(), 1))
+		err := roachtestutil.WaitFor3XReplication(ctx, t, t.L(), c.Conn(ctx, t.L(), 1))
 		require.NoError(t, err)
 
 		gossipNetworkAccordingTo := func(node int) (nodes []int) {
