+++ pkg/cmd/roachtest/tests/inconsistency.go
@@ -17,6 +17,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/registry"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
 	"github.com/stretchr/testify/require"
@@ -45,7 +46,7 @@ func runInconsistency(ctx context.Context, t test.Test, c cluster.Cluster) {
 		// to expect it.
 		_, err := db.ExecContext(ctx, `SET CLUSTER SETTING server.consistency_check.interval = '0'`)
 		require.NoError(t, err)
-		require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), db))
+		require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), db))
 		require.NoError(t, db.Close())
 	}
