+++ pkg/cmd/roachtest/tests/jobs_util.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
@@ -93,7 +94,7 @@ func executeNodeShutdown(
 		// is in a healthy state before we start bringing any
 		// nodes down.
 		t.Status("waiting for cluster to be 3x replicated")
-		err := WaitFor3XReplication(ctx, t, t.L(), watcherDB)
+		err := roachtestutil.WaitFor3XReplication(ctx, t, t.L(), watcherDB)
 		if err != nil {
 			return err
 		}
