+++ pkg/cmd/roachtest/tests/limit_capacity.go
@@ -85,7 +85,7 @@ func runLimitCapacity(ctx context.Context, t test.Test, c cluster.Cluster, cfg l
 	conn := c.Conn(ctx, t.L(), 1)
 	defer conn.Close()
 
-	require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+	require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 	var cancels []func()
 
 	c.Run(ctx, option.WithNodes(c.WorkloadNode()), "./cockroach workload init kv --splits=1000 {pgurl:1}")
