+++ pkg/cmd/roachtest/tests/many_splits.go
@@ -16,6 +16,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
 	"github.com/stretchr/testify/require"
@@ -32,7 +33,7 @@ func runManySplits(ctx context.Context, t test.Test, c cluster.Cluster) {
 	defer db.Close()
 
 	// Wait for up-replication then create many ranges.
-	err := WaitFor3XReplication(ctx, t, t.L(), db)
+	err := roachtestutil.WaitFor3XReplication(ctx, t, t.L(), db)
 	require.NoError(t, err)
 
 	m := c.NewMonitor(ctx, c.All())
