+++ pkg/cmd/roachtest/tests/mismatched_locality.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
 )
@@ -58,7 +59,7 @@ func runMismatchedLocalityTest(ctx context.Context, t test.Test, c cluster.Clust
 	// Verify that we can add and drop regions for the database. There's no longer
 	// any node with the old localities, but that's fine.
 	db = c.Conn(ctx, t.L(), 3)
-	if err := WaitFor3XReplication(ctx, t, t.L(), db); err != nil {
+	if err := roachtestutil.WaitFor3XReplication(ctx, t, t.L(), db); err != nil {
 		t.Fatal(err)
 	}
 	if _, err := db.Exec(`ALTER DATABASE defaultdb ADD REGION "venus";`); err != nil {
