+++ pkg/cmd/roachtest/tests/multi_region_system_database.go
@@ -19,6 +19,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/registry"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/spec"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
@@ -47,7 +48,7 @@ func registerMultiRegionSystemDatabase(r registry.Registry) {
 			conn := c.Conn(ctx, t.L(), 1)
 			defer conn.Close()
 
-			require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+			require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 			_, err := conn.ExecContext(ctx, "SET CLUSTER SETTING sql.multiregion.system_database_multiregion.enabled = true")
 			require.NoError(t, err)
@@ -111,7 +112,7 @@ func registerMultiRegionSystemDatabase(r registry.Registry) {
 			}
 
 			for i := 2; i <= nodes; i++ {
-				require.NoError(t, WaitFor3XReplication(ctx, t, t.L(), conn))
+				require.NoError(t, roachtestutil.WaitFor3XReplication(ctx, t, t.L(), conn))
 
 				t.WorkerStatus("stop")
 				c.Run(ctx, option.WithNodes(c.Node(i)), "killall -9 cockroach")
